Move header files to scr
