+++ src/Makevars
@@ -1,4 +1,4 @@
-PKG_CPPFLAGS = -I../inst/include/ -I../inst/include/ezc3d/
+PKG_CPPFLAGS = -I.
 
 ##include files manually
 C3DRSOURCES = \
