+++ src/ezc3d/AnalogsInfo.h
