+++ src/ezc3d/Channel.h
