+++ src/ezc3d/RotationsInfo.h
