+++ src/read.cpp
@@ -1,6 +1,6 @@
 #include <Rcpp.h>
-#include "ezc3d_all.h"
-#include "matrix_conversion.h"
+#include "ezc3d/ezc3d_all.h"
+#include "ezc3d/matrix_conversion.h"
 
 // [[Rcpp::export]]
 Rcpp::List read(const std::string &filepath) {
